Move PlayerState in the right folder
